rename range_map to segments log cleanup
