+++ crates/fluvio-storage/src/segments.rs
